requirements.txt
PySide6
Pillow
fpdf2
